refactor(sqlgen): split benchmark projections out of duckdb_schema_projection.go (#220 seam, #260 review) Signed-off-by: Ruoshi Li <iceboundrock@gmail.com>
